Fixed tests. Client search accepts either objects or ids

+++ lib/gooddata/models/domain.rb
@@ -423,7 +423,7 @@ module GoodData
       data = GoodData::Helpers.get_path(res, ['updateClientsResponse'])
       if data
         result = data.flat_map { |k, v| v.map { |h| GoodData::Helpers.symbolize_keys(h.merge('type' => k)) } }
-        result.select { |r| r[:status] == 'DELETED' }.peach { |r| r[:originalProject] && client.delete(r[:originalProject]) }
+        result.select { |r| r[:status] == 'DELETED' }.peach { |r| r[:originalProject] && client.delete(r[:originalProject]) } if options[:delete_projects]
         result
       else
         []
